scripts/unused/requirements.txt
appdirs==1.4.3
gspread==0.6.2
httplib2==0.10.3
oauth2client==4.0.0
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
rsa==3.4.2
six==1.10.0
